Provide routes as expected by angular ui-router
AngularUI's UI-Router has a state mechanism that allows you to name states
that watch URLs and act according to them. Look at the angular-symfony-demo
repository for more details on how to do that.

To avoid repetitiveness in the url routing, this library allows you to use
the named routes defined in symfony and not repeat that again in the javascript
code that angularjs looks up.